llm/client.py
import os
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()

def get_openai_client():
    """
    Returns an authenticated OpenAI client.
    Raises ValueError if OPENAI_API_KEY is not set.
    """
    api_key = os.getenv("OPENAI_API_KEY")
    if not api_key:
        raise ValueError("OPENAI_API_KEY not found in environment variables.")
    return OpenAI(api_key=api_key)
