requirements.txt
pydantic
openai
requests
python-dotenv
fastapi
uvicorn
jinja2
